+++ bobtemplates/plone/theme_barceloneta.py
@@ -1,18 +1,15 @@
 # -*- coding: utf-8 -*-
 
-import os
-import re
-
+from bobtemplates.plone.base import base_prepare_renderer
+from bobtemplates.plone.base import echo
+from bobtemplates.plone.base import git_commit
+from bobtemplates.plone.base import update_file
+from bobtemplates.plone.base import validate_packagename
 from lxml import etree
 from mrbob.bobexceptions import ValidationError
 
-from bobtemplates.plone.base import (
-    base_prepare_renderer,
-    echo,
-    git_commit,
-    update_file,
-    validate_packagename,
-)
+import os
+import re
 
 
 def pre_theme_name(configurator, question):