+++ package_tests/test_portlet.py
@@ -2,13 +2,13 @@
 
 """Test view generation."""
 
-import os
-
-import pytest
+from bobtemplates.plone import base
+from bobtemplates.plone import portlet
 from mrbob.bobexceptions import ValidationError
 from mrbob.configurator import Configurator
 
-from bobtemplates.plone import base, portlet
+import os
+import pytest
 
 
 def test_pre_renderer():