+++ skeleton-tests/base.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
 
 # from bobtemplates.plone.utils import safe_encode
+from bobtemplates.plone.utils import safe_unicode
+
 import contextlib
 import logging
 import os
 import subprocess
 
-from bobtemplates.plone.utils import safe_unicode
 
 logger = logging.getLogger("skeleton-tests")
 
@@ -27,6 +28,16 @@ def file_exists(base_path, file_path):
 
 
 def run_skeleton_tox_env(wd, config):
+    logger.info("show_versions: python / tox")
+    print("show_versions: python / tox")
+    subprocess.check_call(
+        ["python", "--version"],
+        cwd=wd,
+    )
+    subprocess.check_call(
+        ["tox", "--version"],
+        cwd=wd,
+    )
     try:
         returncode = subprocess.check_call(
             ["tox", "-e", config.skeleton_tox_env, "-p", "auto", "-o", "-v", "-r"],